Google Data Studio Version
Everything came out great, BUT the map isn't rendering properly due to an error that needs to be remedied by the service center. Otherwise, I'm content with how this turned out and it was so easy! I recommend Google Data Studio for those who want an easy data visualization tool, can create calculated fields using familiar Excel function syntax, and do not want to, or cannot, pay for a data visualization tool for their personal use.